+++ src/client/widgets/AutoUpdateDialog.cpp
@@ -136,7 +136,7 @@ void AutoUpdateDialog::checkForUpdate() {
 		resultMsg.setWindowFlags(resultMsg.windowFlags() & ~Qt::WindowContextHelpButtonHint);
 		resultMsg.exec();
 
-		const int result = reinterpret_cast<int>(::ShellExecuteA(nullptr, "runas", exeFileName.toUtf8().constData(), nullptr, nullptr, SW_SHOWNORMAL));
+		const int64_t result = reinterpret_cast<int64_t>(::ShellExecuteA(nullptr, "runas", exeFileName.toUtf8().constData(), nullptr, nullptr, SW_SHOWNORMAL));
 		if (result > 32) { // no error
 			qApp->quit();
 		}
